Copy gcs to bq
